requirements.txt
selenium==4.12.0
termcolor==2.3.0
azure-identity==1.14.0
pandas==2.1.0